// frontend/src/features/Restaurants/types.ts
export interface Restaurant {
    name: string;
    phoneNumber: string;
    mobilePhoneNumber: string;
    email: string;
    city: string;
    district: string;
};
export interface CreateRestaurant {
    name: string;
    phoneNumber: string;
    mobilePhoneNumber: string;
    email: string;
    city: string;
    district: string;
};